test: move relevant test from createImageWizard to targetEnvironment test file
this commit move relvant test from creatreImageWizrad to targetEnvironment test file
